requirements.txt
ttkbootstrap>=1.10.0
pyyaml>=6.0
requests>=2.28.0
pypresence>=4.3.0
Pillow>=9.0.0
pyinstaller>=5.0.0
